requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic[email]
python-decouple
aio-pika
passlib==5.3.0
bcrypt==4.1.2
